Organize assets into folder structure
