+++ assets/css/styles.css
@@ -29,7 +29,7 @@ body {
 
 /* Parallax background child */
 .parallax-bg {
-  background-image: url("background.jpg");
+  background-image: url("../images/background.jpg");
   background-repeat: repeat-y;
   background-size: 100% auto;
   background-position: center top;
@@ -40,7 +40,7 @@ body {
 
 /* Parallax foreground child with blend mode */
 .parallax-fg {
-  background-image: url("Layer 5.jpg");
+  background-image: url("../images/layer-5.jpg");
   background-repeat: repeat-y;
   background-size: 100% auto;
   background-position: center top;
@@ -62,7 +62,7 @@ body {
   left: 0;
   width: 100%;
   height: 100%;
-  background-image: url("overlay.jpg");
+  background-image: url("../images/overlay.jpg");
   background-repeat: repeat;
   background-position: center center;
   background-size: 720px auto;
