+++ index.html
@@ -4,14 +4,14 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <title>nepobabiesruntheunderground</title>
-    <link rel="stylesheet" href="styles.css" />
+    <link rel="stylesheet" href="assets/css/styles.css" />
     <meta
       name="theme-color"
       content="#ffffff"
       media="(prefers-color-scheme: dark)"
     />
-  <link rel="icon" type="image/png" href="favicon.png" />
-  <script defer src="parallax.js"></script>
+  <link rel="icon" type="image/png" href="assets/images/favicon.png" />
+  <script defer src="assets/js/parallax.js"></script>
   </head>
   <body>
     <div id="background">
@@ -41,12 +41,12 @@
       <a href="https://www.instagram.com/nepobabiesruntheunderground/">
         <div class="bassvictim-container">
           <img
-            src="bassvictim.png"
+            src="assets/images/bassvictim.png"
             class="background_bassvictim"
             alt="Makes You Wonder"
           />
           <img
-            src="bassvictim_overlay.png"
+            src="assets/images/bassvictim_overlay.png"
             class="overlay_bassvictim"
             alt="Makes You Wonder"
           />
@@ -67,33 +67,33 @@
     </section>
 
     <section id="nettspend-parallax">
-      <img src="nettspend.gif" alt="" class="drankgif drankdrankdrank" />
-      <img src="nettspend_overlay.png" class="drankdrankdrank" />
+      <img src="assets/images/nettspend.gif" alt="" class="drankgif drankdrankdrank" />
+      <img src="assets/images/nettspend_overlay.png" class="drankdrankdrank" />
       <a href="https://www.last.fm/user/TheTexta">
-        <img src="nettspend_overlay.png" class="drankhue drankdrankdrank" />
+        <img src="assets/images/nettspend_overlay.png" class="drankhue drankdrankdrank" />
       </a>
     </section>
 
     
     <section>
       <a href="./me.html">
-        <img src="me_background.gif" alt="" class="me_background me" loading="lazy" />
-        <img src="me_foreground.gif" alt="" class="me_foreground me" loading="lazy" />
-        <img src="me_background.gif" alt="" class="me_ultraforeground me" loading="lazy" />
-        <img src="me_hover.gif" alt="" class="me_hover me" loading="lazy" />
+        <img src="assets/images/me_background.gif" alt="" class="me_background me" loading="lazy" />
+        <img src="assets/images/me_foreground.gif" alt="" class="me_foreground me" loading="lazy" />
+        <img src="assets/images/me_background.gif" alt="" class="me_ultraforeground me" loading="lazy" />
+        <img src="assets/images/me_hover.gif" alt="" class="me_hover me" loading="lazy" />
       </a>
     </section>
 
     <section>
       <a href="https://open.spotify.com/user/djaydex?si=b02c31c1b8a248ca">
         <img
-          src="pilleater.png"
+          src="assets/images/pilleater.png"
           alt=""
           class="pilleater-image"
           id="pilleater-background"
         />
         <img
-          src="pilleater.png"
+          src="assets/images/pilleater.png"
           alt=""
           class="pilleater-image-foreground pilleater-image"
           id="pilleater-foreground"
