+++ journal.json
@@ -3,7 +3,7 @@
     "title": "Entree #1",
     "date": "2025-07-15",
     "quote of the day": "good taste is careful not cheap tiktok dopamine",
-    "content": "im going to write here \n <img src=\"blog photos/entree one photo one.png\"> \n i dont know what to write \n im not going to edit entrees once the day has passed - whatever time zone i may be in"
+    "content": "im going to write here \n <img src=\"assets/images/blog/entree-one-photo-one.png\"> \n i dont know what to write \n im not going to edit entrees once the day has passed - whatever time zone i may be in"
   },
   {
     "title": "Entree #2",
