+++ me.html
@@ -5,8 +5,8 @@
   <title>me</title>
   <meta charset="UTF-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1" />
-  <link rel="stylesheet" href="styles.css" />
-  <link rel="icon" type="image/png" href="favicon.png" />
+  <link rel="stylesheet" href="assets/css/styles.css" />
+  <link rel="icon" type="image/png" href="assets/images/favicon.png" />
 </head>
 
 <body>
@@ -69,6 +69,6 @@
   }
 
   </script>
-  <script src="formatDate.js"></script>
+  <script src="assets/js/formatDate.js"></script>
   </body>
   </html>
